+++ R/fetch_sigtab.R
@@ -24,16 +24,21 @@ fetch_sigtab <- function(timeout = 240) {
   sigtab_url <- "ftp://ftp.datasus.gov.br/dissemin/publicos/SIASUS/200801_/Auxiliar/TAB_SIA.zip"
   sigtab_file <- file.path("DBF/TB_SIGTAW.dbf")
 
-  cli::cli_alert_info("Downloading the current SIGTAB table from DataSUS...")
+  cli::cli_alert_info(
+    "Downloading DataSUS auxiliary table {.strong SIGTAB}..."
+  )
   tmp <- .datasus_fetch_zip_dbf(sigtab_url, sigtab_file, timeout)
   if (ncol(tmp) != 2L) {
     cli::cli_abort(
-      "SIGTAB must contain exactly two columns; found {ncol(tmp)}."
+      "The downloaded SIGTAB table must contain exactly two columns; found {ncol(tmp)}."
     )
   }
   colnames(tmp) <- c("COD", "nome_proced")
   tmp$COD <- as.character(tmp$COD)
   tmp$nome_proced <- stringi::stri_enc_toutf8(tmp$nome_proced)
 
+  cli::cli_alert_success(
+    "Downloaded and read DataSUS auxiliary table {.strong SIGTAB}."
+  )
   tmp
 }